feat(assessment): Hexagonal Architecture 전환 (#29)
- Entity를 adapter/out/persistence/entity로 이동
- Repository를 JpaRepository로 리네이밍
- Controller를 adapter/in/web으로 이동
- DTOs를 adapter/in/web/dto로 이동
- Service를 application/service로 이동
- Util을 application/util로 이동
- 타 도메인 import 경로 업데이트
