+++ springboot/src/main/java/com/mzc/backend/lms/domains/assessment/adapter/in/web/dto/response/AssessmentDetailResponseDto.java
@@ -1,7 +1,7 @@
-package com.mzc.backend.lms.domains.assessment.dto.response;
+package com.mzc.backend.lms.domains.assessment.adapter.in.web.dto.response;
 
-import com.mzc.backend.lms.domains.assessment.entity.Assessment;
-import com.mzc.backend.lms.domains.assessment.enums.AssessmentType;
+import com.mzc.backend.lms.domains.assessment.adapter.out.persistence.entity.Assessment;
+import com.mzc.backend.lms.domains.assessment.adapter.out.persistence.entity.enums.AssessmentType;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
