+++ springboot/src/main/java/com/mzc/backend/lms/domains/assessment/adapter/in/web/dto/response/ProfessorAttemptListItemResponseDto.java
@@ -1,6 +1,5 @@
-package com.mzc.backend.lms.domains.assessment.dto.response;
+package com.mzc.backend.lms.domains.assessment.adapter.in.web.dto.response;
 
-import com.fasterxml.jackson.databind.JsonNode;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -11,13 +10,13 @@ import java.time.LocalDateTime;
 @Getter
 @Builder
 @AllArgsConstructor
-public class ProfessorAttemptDetailResponseDto {
+public class ProfessorAttemptListItemResponseDto {
 
     private Long attemptId;
     private Long examId;
     private Long courseId;
 
-    private ProfessorAttemptListItemResponseDto.StudentInfo student;
+    private StudentInfo student;
 
     private LocalDateTime startedAt;
     private LocalDateTime submittedAt;
@@ -28,16 +27,14 @@ public class ProfessorAttemptDetailResponseDto {
     private BigDecimal score;
     private String feedback;
 
-    /**
-     * 명세서 형식에 맞추기 위해 래핑:
-     * { "answers": { "q1": 1, ... } }
-     */
-    private JsonNode answerData;
-
-    /**
-     * 교수 조회는 정답 포함 원본 제공
-     */
-    private JsonNode questionData;
+    @Getter
+    @Builder
+    @AllArgsConstructor
+    public static class StudentInfo {
+        private Long id;
+        private String studentNumber;
+        private String name;
+    }
 }
